+++ module-auth/src/main/java/exam/demo/moduleauth/pojo/model/Role.java
@@ -13,8 +13,11 @@ import java.util.Set;
 /**
  * 权限表 - 数据对象定义
  *
- * @author gpmscloud
+ * @author luohui
+ * @version 1.0
+ * @since 2020-03-04
  */
+
 @Data
 @TableName("role")
 public class Role implements Serializable {
@@ -23,17 +26,17 @@ public class Role implements Serializable {
      * 权限id
      */
     @ApiModelProperty(value = "权限id")
-    private Integer id;
+    private Long id;
     /**
      * 公司id
      */
     @ApiModelProperty(value = "公司id")
-    private Integer companyId;
+    private Long companyId;
     /**
      * 组织机构id
      */
     @ApiModelProperty(value = "组织机构id")
-    private Integer orgId;
+    private Long orgId;
     /**
      * 角色名
      */
@@ -53,12 +56,12 @@ public class Role implements Serializable {
      * 状态位
      */
     @ApiModelProperty(value = "状态位")
-    private Integer status;
+    private Long status;
     /**
      * 创建人
      */
     @ApiModelProperty(value = "创建人")
-    private Integer createdBy;
+    private Long createdBy;
     /**
      * 创建时间
      */
@@ -70,7 +73,7 @@ public class Role implements Serializable {
      * 修改人
      */
     @ApiModelProperty(value = "修改人")
-    private Integer updatedBy;
+    private Long updatedBy;
     /**
      * 修改时间
      */
