+++ module-auth/src/main/java/exam/demo/moduleauth/pojo/model/UserInfo.java
@@ -9,11 +9,11 @@ import lombok.experimental.Accessors;
 @Accessors(chain = true)
 public class UserInfo {
     @JsonSerialize(using = ToStringSerializer.class)
-    private Integer id;
+    private Long id;
     private String name;
     private String profilePicture;
     @JsonSerialize(using = ToStringSerializer.class)
-    private Integer companyId;
+    private Long companyId;
 
     public String getName() {
         return name;
@@ -31,19 +31,19 @@ public class UserInfo {
         this.profilePicture = profilePicture;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public Integer getCompanyId() {
+    public Long getCompanyId() {
         return companyId;
     }
 
-    public void setCompanyId(Integer companyId) {
+    public void setCompanyId(Long companyId) {
         this.companyId = companyId;
     }
