+++ module-common/src/main/java/exam/demo/modulecommon/common/FuzzySearch.java
@@ -18,5 +18,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class FuzzySearch {
     String paperName;
-    Integer companyId;
+    Long companyId;
 }
