+++ module-common/src/main/java/exam/demo/modulecommon/common/PaperSubjectDto.java
@@ -23,28 +23,28 @@ public class PaperSubjectDto {
      * 试题id
      */
     @JsonSerialize(using = ToStringSerializer.class)
-    private Integer id;
+    private Long id;
     /**
      * 试卷id
      */
     @JsonSerialize(using = ToStringSerializer.class)
-    private Integer paperId;
+    private Long paperId;
     /**
      * 题目
      */
     private String subject;
     @JsonSerialize(using = ToStringSerializer.class)
-    private Integer categoryId;
+    private Long categoryId;
     /**
      * 难度
      */
     @JsonSerialize(using = ToStringSerializer.class)
-    private Integer difficulty;
+    private Long difficulty;
     /**
      * 题目类型从基础数据中取
      */
     @JsonSerialize(using = ToStringSerializer.class)
-    private Integer subjectTypeId;
+    private Long subjectTypeId;
     /**
      * 是否为客观题
      */
