+++ module-exam/src/main/java/exam/demo/moduleexam/pojo/DTO/grade/ExamGradeRecordTableDataDTO.java
@@ -6,13 +6,13 @@ import java.util.Date;
 
 @Data
 public class ExamGradeRecordTableDataDTO {
-    private Integer id;
+    private Long id;
     private Date actualEndTime;
-    private Integer examSession;
+    private Long examSession;
     private String examiner;
     private Date markingStopTime;
     private Double objectiveSubjectScore;
-    private Integer paperId;
+    private Long paperId;
     private Date createdTime;
     private Byte status;
     private Double subjectvieSubjectScore;
