+++ module-exam/src/main/java/exam/demo/moduleexam/pojo/DTO/grade/MarkingAnswerDTO.java
@@ -23,5 +23,5 @@ public class MarkingAnswerDTO {
     /**
      * 试题id
      */
-    private Integer paperSubjectId;
+    private Long paperSubjectId;
 }
