+++ module-exam/src/main/java/exam/demo/moduleexam/pojo/DTO/grade/MyAnswerDTO.java
@@ -14,5 +14,5 @@ public class MyAnswerDTO {
     private String myAnswer;
     private String standardAnswer;
     @JsonSerialize(using = ToStringSerializer.class)
-    private Integer paperSubjectId;
+    private Long paperSubjectId;
 }
