+++ module-exam/src/main/java/exam/demo/moduleexam/pojo/DTO/grade/QueryPaperAnswerDTO.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class QueryPaperAnswerDTO {
-    private Integer paperId;
-    private Integer examRecordId;
+    private Long paperId;
+    private Long examRecordId;
 }
