+++ module-exam/src/main/java/exam/demo/moduleexam/pojo/DTO/publish/ExamPublishRecordPublishFormDTO.java
@@ -11,13 +11,13 @@ public class ExamPublishRecordPublishFormDTO extends BaseDto {
     private String descript;
     private Date startTime;
     private Date endTime;
-    private Integer examSession;
-    private Integer limitTime;
-    private Integer markingMode;
+    private Long examSession;
+    private Long limitTime;
+    private Long markingMode;
     private Date markingStopTime;
-    private Integer paperId;
-    private Integer planPepoleNum;
+    private Long paperId;
+    private Long planPepoleNum;
     private String title;
-    private List<Integer> examinersId;
-    private Integer publisher;
+    private List<Long> examinersId;
+    private Long publisher;
 }
