+++ module-exam/src/main/java/exam/demo/moduleexam/pojo/DTO/publish/ExamPublishRecordTableDataDTO.java
@@ -7,18 +7,18 @@ import java.util.List;
 
 @Data
 public class ExamPublishRecordTableDataDTO {
-    private Integer id;
-    private Integer publisher;
+    private Long id;
+    private Long publisher;
     private Date createdTime;
     private Date endTime;
-    private Integer planPepoleNum;
-    private Integer limitTime;
+    private Long planPepoleNum;
+    private Long limitTime;
     private String title;
     private String descript;
     private Byte status;
     private Long version;
-    private Integer examSession;
-    private Integer publishTimes;
-    private List<Integer> examiners;
-    private Integer paperId;
+    private Long examSession;
+    private Long publishTimes;
+    private List<Long> examiners;
+    private Long paperId;
 }
